AccountWithSortCode: Add code and tests
It seems it's common to enter the account number and the sort code
separately. Still the validator requires them both.

I've tried to find a reference to the account number min/max length,
but no luck. So say that they are 8 digits, but some say that they vary
from 6 to 10 digits.